Migrate from Next.js pages router to app router
This migrates the entire site from the pages router to the app router architecture.

Key changes:
- Replaced src/pages/_app.jsx and src/pages/_document.jsx with src/app/layout.jsx
- Migrated all pages to src/app/ directory structure
- Converted getStaticProps to async Server Components
- Converted getStaticPaths to generateStaticParams
- Replaced next/head with metadata exports
- Added 'use client' directive to components using React hooks
- Created EpisodePlayButton client component for episode playback
- Updated Layout component to use usePathname instead of useRouter
- Removed experimental scrollRestoration from next.config.js (built into app router)

All 186 pages build successfully with static export enabled.
