+++ next.config.js
@@ -1,9 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  experimental: {
-    scrollRestoration: true,
-  },
   output: 'export',
   trailingSlash: true,
   images: {
