+++ src/app/pitch/page.jsx
@@ -1,0 +1,19 @@
+import { ProseContainer } from '@/components/ProseContainer'
+
+export const metadata = {
+  title: 'Pitch a Guest - Legacy Code Rocks - Explore the world of modernizing existing software applications',
+  description: 'Explore the world of modernizing existing software applications',
+}
+
+export default function Pitch() {
+  return (
+    <ProseContainer>
+      <h2>Know Someone We Should Talk To?</h2>
+      <p>
+        We&apos;re always looking for new guests to have no the show. If you
+        know who would be a good fit for the show (even if it&apos;s you.),
+        then please fill out <a href="https://9lsmuup7u2y.typeform.com/to/FXr5kFRS">our pitch form</a>.
+      </p>
+    </ProseContainer>
+  )
+}
