+++ src/components/EpisodeEntry.jsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { useMemo } from 'react'
 import Link from 'next/link'
