--- src/pages/404.jsx
@@ -1,27 +1,0 @@
-import Link from 'next/link'
-
-export default function ErrorPage() {
-  return (
-    <>
-      <main className="grid min-h-full place-items-center px-6 py-24 sm:py-32 lg:px-8">
-        <div className="text-center">
-          <p className="text-base font-semibold text-brand-red-500">404</p>
-          <h1 className="mt-4 text-3xl font-bold tracking-tight text-gray-900 sm:text-5xl">Page not found</h1>
-          <p className="mt-6 text-base leading-7 text-gray-600">Sorry, we couldn’t find the page you’re looking for.</p>
-          <div className="mt-10 flex items-center justify-center gap-x-6">
-            <Link
-              href="/"
-              className="rounded-md bg-brand-red-500 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-brand-red-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-brand-red-600"
-            >
-              Go back home
-            </Link>
-          </div>
-        </div>
-      </main>
-    </>
-  )
-}
-
-ErrorPage.getLayout = function getLayout(page) {
-  return page
-}
